requirements.txt
asgiref==3.8.1
Django==5.1.1
django-environ==0.11.2
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.1